core: Teardown links before calling event handlers
Previously, the link up/down event handlers were called in such a way that if
you sent a message from inside the handler, it at least might go through.  Now
links should be torn down all the way before handlers are called, and it should
not be possible to send messages from them.  This makes much more sense.

Additionally, exceptions in link event handlers used to potentially cause
problems for the simulator, but should now be relatively safe.

Finally, there's part of a tweak here to make the whole simulator a bit more
robust with respect to module load order.